requirements.txt
Django>=2.1
mysqlclient>=1.4.4
djangorestframework>=3.1
djangorestframework-xml>=1.4
